OGM-1152 Move the dialect invocations tests to a specific package
