+++ core/src/test/java/org/hibernate/ogm/backendtck/dialectinvocations/GridDialectOperationInvocationForElementCollectionTest.java
@@ -11,7 +11,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.backendtck.embeddable;
+package org.hibernate.ogm.backendtck.dialectinvocations;
 
 import java.util.List;
 import java.util.Map;
